requirements.txt
discord.py
feedparser
python-dateutil
Flask
